+++ main.go
@@ -14,7 +14,7 @@ import (
 
 var (
 	// Version is the main version number (injected by GoReleaser via -ldflags)
-	Version = "0.1.0"
+	Version = "0.5.0"
 
 	// VersionPrerelease is a pre-release marker for the version (injected by GoReleaser)
 	VersionPrerelease = "dev"